RavenDB-9267 Dashboard disconnects every 2 seconds

+++ src/Raven.Server/Dashboard/MachineResourcesNotificationSender.cs
@@ -75,8 +75,8 @@ namespace Raven.Server.Dashboard
                 TotalMemory = installedMemory,
                 MachineMemoryUsage = installedMemory - availableMemory,
                 ProcessMemoryUsage = workingSet,
-                MachineCpuUsage = cpuInfo.MachineCpuUsage,
-                ProcessCpuUsage = cpuInfo.ProcessCpuUsage
+                MachineCpuUsage = (int)cpuInfo.MachineCpuUsage,
+                ProcessCpuUsage = (int)cpuInfo.ProcessCpuUsage
             };
 
             return machineResources;
